requirements.txt
python-telegram-bot[webhooks]>=21.0
python-dotenv>=1.0.0
openai>=1.0.0
supabase>=2.0.0
